+++ console/cypress/integration/api-explorer/graphql/test.ts
@@ -16,7 +16,6 @@ const setup = () => {
     it('Visit the index route', () => {
       // Visit the index route
       cy.visit('/');
-      cy.wait(7000);
       // Get and set validation metadata
       setMetaData();
     });