+++ console/src/dataSources/services/postgresql/__tests__/postgresql.test.ts
@@ -30,4 +30,79 @@ describe('postgresql datasource tests', () => {
       expect(query).toMatchSnapshot();
     });
   });
+
+  describe('getCreateIndexSQL', () => {
+    const { createIndexSql } = postgres;
+    it('generates proper SQL for column names containing no spaces', () => {
+      if (createIndexSql) {
+        const query = createIndexSql({
+          table: { schema: 'test_schema', name: 'test_table' },
+          columns: ['firstname', 'lastname'],
+          indexName: 'test_index_name',
+          indexType: 'btree',
+          unique: false,
+        });
+        expect(query).toMatchInlineSnapshot(`
+          "
+            CREATE  INDEX \\"test_index_name\\" on
+            \\"test_schema\\".\\"test_table\\" using btree (\\"firstname\\", \\"lastname\\");
+          "
+        `);
+      }
+    });
+
+    it('generates proper SQL for column names in camel case', () => {
+      if (createIndexSql) {
+        const query = createIndexSql({
+          table: { schema: 'test_schema', name: 'test_table' },
+          columns: ['firstName', 'lastName'],
+          indexName: 'test_index_name',
+          indexType: 'btree',
+          unique: false,
+        });
+        expect(query).toMatchInlineSnapshot(`
+          "
+            CREATE  INDEX \\"test_index_name\\" on
+            \\"test_schema\\".\\"test_table\\" using btree (\\"firstName\\", \\"lastName\\");
+          "
+        `);
+      }
+    });
+
+    it('generates proper SQL for column names containing spaces', () => {
+      if (createIndexSql) {
+        const query = createIndexSql({
+          table: { schema: 'test_schema', name: 'test_table' },
+          columns: ['first name', 'last name'],
+          indexName: 'test_index_name',
+          indexType: 'btree',
+          unique: false,
+        });
+        expect(query).toMatchInlineSnapshot(`
+          "
+            CREATE  INDEX \\"test_index_name\\" on
+            \\"test_schema\\".\\"test_table\\" using btree (\\"first name\\", \\"last name\\");
+          "
+        `);
+      }
+    });
+
+    it('generates proper SQL for with unique param set to true', () => {
+      if (createIndexSql) {
+        const query = createIndexSql({
+          table: { schema: 'test_schema', name: 'test_table' },
+          columns: ['first name', 'last_name', 'fullName'],
+          indexName: 'test_index_name',
+          indexType: 'btree',
+          unique: true,
+        });
+        expect(query).toMatchInlineSnapshot(`
+          "
+            CREATE UNIQUE INDEX \\"test_index_name\\" on
+            \\"test_schema\\".\\"test_table\\" using btree (\\"first name\\", \\"last_name\\", \\"fullName\\");
+          "
+        `);
+      }
+    });
+  });
 });