+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - server: fix GraphQL type for remote relationship field (close #7284)
 - server: support EdDSA algorithm and key type for JWT
 - server: fix GraphQL type for single-row returning functions (close #7109)
+- server: update non-existent event trigger, action and query collection error msgs (close #7396) 
 - console: add support for creation of indexes for Postgres data sources
 - console: allow same named queries and unnamed queries on allowlist file upload
 - console: support computed fields in permission builder