+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - server: fix for incorrect `__typename` value in nested remote joins with a customized remote schema
 - server: fix a bug where some unicode characters in default string values for fields in remote schemas could lead to internal errors
 - server: bigquery: implement `_in` and `_nin` operators. (close #7343)
+- server: bigquery: custom root names, table names and field names for bigquery are included in tests
 - console: fix untracked foreign-key relationships suggestion across schemas
 - console: allow resolution of conflicting inherited role permissions
 - cli: fix SDL formatting in `actions.graphql`(#7296)